% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{acc_fetch}
\alias{acc_fetch}
\title{acc_fetch}
\usage{
acc_fetch(db, table)
}
\arguments{
\item{db}{Required. Database location and name}

\item{table}{Required. Name of the table}
}
\value{
Chosen table from database
}
\description{
fetches table from Access database then closes connection
}
\examples{
\dontrun{

library(RODBC)

table <- acc_fetch(db = "D:/my_database.accdb", table = "my_table")

}
}

