% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{clean_corpus}
\alias{clean_corpus}
\title{clean_corpus}
\usage{
clean_corpus(corpus)
clean_corpus(corpus, stem=FALSE)
clean_corpus(corpus, stem=FALSE, rem_words="the")
}
\arguments{
\item{corpus}{Required. Corpus of documents to clean.}

\item{stem}{Optional, boolean. If TRUE stems documents after cleaning. Defaults to FALSE.}

\item{rem_words}{Optional, additional words to remove from documents.}
}
\value{
cleaned corpus
}
\description{
Cleans corpus of documents.
}
\details{
Uses tm_map functions from \href{http://cran.r-project.org/web/packages/tm/tm.pdf}{tm package}.

* Strips white space
* To lower case
* Removes @tags
* Removes punctuation
* Removes numbers
* Removes stopwords
* To plain text
}
\examples{
\dontrun{

 library(tm)

 data(crude)

 crude <- clean_corpus(crude, stem = TRUE, rem_words = c("saudi"))

 }
}
\seealso{
\href{http://cran.r-project.org/web/packages/tm/tm.pdf}{tm package}
}

