% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{pam_k}
\alias{pam_k}
\title{pam_k}
\usage{
pam_k(tdm)
pam_k(tdm, sparsity = 0.98)
}
\arguments{
\item{tdm}{Required. object of class TermDocumentMatrix.}

\item{sparsity}{Optional. Maximum allowed sparsity. Between 0 and 1, defaults to 0.98.}
}
\value{
2D clustering plot - clusters
Silhouette plot - higher the si the better clustered
Prints clusters in console
}
\description{
Like k-means, attempts to break dataset into groups and minimize distance between points.
}
\details{
Partitions around \href{http://en.wikipedia.org/wiki/K-medoids}{k-medoids} and produces 2D clustering plot as well as silhouette plot to test the fit of clusters.
}
\examples{
\dontrun{
library(tm)

data(crude)

tdm <- TermDocumentMatrix(crude, control=list(minWordLength=1))

pam_k(tdm, sparsity = 0.99)
}
}
\seealso{
\code{\link{hier_clust}} for hierarchical clustering
\href{http://cran.r-project.org/web/packages/fpc/fpc.pdf}{fpc} package (dependency)
}

