% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{plot_topic}
\alias{plot_topic}
\title{plot_topic}
\usage{
plot_topic(tdm, date, topics = 5, terms = 4)
}
\arguments{
\item{tdm}{Required. term-document matri; must be of class "TermDocumentMatrix"}

\item{date}{Required. Vector of dates to plot topics against. POSIXlt class}

\item{topics}{Optional. Number of topics to find (numeric). Defaults to 5.}

\item{terms}{Optional. Number of terms to display for each topics; defaults to 4.}
}
\value{
Plots of topics over time (ggplot2)
}
\description{
Models topics using Latent Dirichlet Allocation and plots against time using qplot.
}
\details{
Uses \href{http://cran.r-project.org/web/packages/topicmodels/topicmodels.pdf}{topicmodels} package as well as \href{http://ggplot2.org/}{ggplot2} package to plot topics over time.
Topics are found with the LDA model - \href{http://en.wikipedia.org/wiki/Latent_Dirichlet_allocation}{Latent Dirichlet Allocation}.
}
\examples{
\dontrun{

library(tm)

data(crude)

tdm <- TermDocumentMatrix(crude, control=list(minWordLength=1))
x <- as.POSIXlt(c(Sys.Date() - 1, Sys.Date(), Sys.Date() + 1 ))
dates <- sample(x, 20, replace =TRUE)

plot_topic(tdm, dates)

}
}

