% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{tag_cloud}
\alias{tag_cloud}
\title{tag_cloud}
\usage{
tag_cloud(tdm)
tag_cloud(tdm, min_freq = 1)
tag_cloud(tdm, min_freq = 30, scale = c(4, .5))
tag_cloud(tdm, min_freq = 150, scale = c(2, .3), order = FALSE)
tag_cloud(tdm, min_freq = 150, scale = c(2, .3), order = FALSE, color = brewer.pal(8, 'Dark2'))
}
\arguments{
\item{tdm}{Required. Term-document matrix}

\item{min_freq}{Optional, defaults to 1. Minimum frequency of term for it to appear in Wordcloud}

\item{scale}{Optional. A vector of length 2 indicating the range of the size of the words.}

\item{order}{Optional, defaults to FALSE. Whether the order of the terms in wordcloud are random}

\item{colour}{Optional. Brewer colour palette, defaults to brewer.pal(8, 'dark2')}
}
\description{
creates word-cloud of most frequent terms in term-document matrix fed to the function
}
\examples{
\dontrun{

library(tm)
library(wordcloud)

data(crude)

tdm <- TermDocumentMatrix(crude, control=list(minWordLength=1))

tag_cloud(tdm, min_freq = 15, scale = c(3, .6), order = FALSE, colour = brewer.pal(9,"BuGn"))

}
}

