% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{term_bar}
\alias{term_bar}
\title{term_bar}
\usage{
term_bar(tdm)
term_bar(tdm, term_freq = 1)
term_bar(tdm, term_freq = 1, colour = "heat")
}
\arguments{
\item{tdm}{Required. Object of class TermDocumentMatrix.}

\item{term_freq}{Optional, defaults to 1. Minimum frequency of terms to plot}

\item{colour}{Optional, defaults to heat. Color to be used for bars.}
}
\description{
term frequency using tdm - barplot
}
\examples{
\dontrun{

library(tm)

data(crude)

tdm <- TermDocumentMatrix(crude, control=list(minWordLength=1))

term_bar(tdm, min.freq=15)

}
}

