% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{term_network}
\alias{term_network}
\title{term_network}
\usage{
term_network(tdm)
term_network(tdm, sparsity = 0.98)
term_network(tdm, sparsity = 0.98, weighted = TRUE)
term_network(tdm, sparsity = 0.98, weighted = TRUE, graphml = FALSE)
}
\arguments{
\item{tdm}{Required. Term-document matrix.}

\item{sparsity}{Optional. Maximum allowed sparsity. Between 0 and 1, defaults to 0.97.}

\item{weighted}{Optional boolean. Whether the graph should be weighted. Defaults to TRUE.}

\item{graphml}{Optional. If TRUE will write graph as graphml file in wd, defaults to FALSE.}
}
\description{
Creates netowrk of terms object (igraph) from term-document matrix
}
\examples{
\dontrun{

 library(tm)

 data(crude)

 tdm <- TermDocumentMatrix(crude, control=list(minWordLength=1))

 net <- term_network(tdm, sparsity = 0.99, weighted=FALSE)

 tkplot(net)

 }
}
\seealso{
\href{http://cran.r-project.org/web/packages/igraph/igraph.pdf}{igraph}
}

